requirements.txt
# Core dependencies
numpy>=1.26.0
streamlit>=1.51.0
sentence-transformers>=5.1.2
google-generativeai>=0.8.5
requests>=2.32.5
torch>=2.9.1

# API dependencies
fastapi>=0.104.0
pydantic>=2.5.0
uvicorn>=0.24.0

# Code conversion
openai>=2.8.1